Fixed writing correct headers in zip archives

+++ src/SharpCompress/Common/Zip/Headers/LocalEntryHeader.cs
@@ -55,8 +55,8 @@ namespace SharpCompress.Common.Zip.Headers
             writer.Write(LastModifiedTime);
             writer.Write(LastModifiedDate);
             writer.Write(Crc);
-            writer.Write(CompressedSize);
-            writer.Write(UncompressedSize);
+			writer.Write((uint)CompressedSize);
+			writer.Write((uint)UncompressedSize);
 
             byte[] nameBytes = EncodeString(Name);
